# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jeevan Voice API - Voice AI agent for low-literacy users
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
  qdrant?: string;
  openai?: string;
}

export type VapiWebhookPayloadMessageCall = {
  id?: string;
};

export type VapiWebhookPayloadMessageCustomer = {
  number?: string;
};

export type VapiWebhookPayloadMessage = {
  type: string;
  call?: VapiWebhookPayloadMessageCall;
  transcript?: string;
  customer?: VapiWebhookPayloadMessageCustomer;
};

export interface VapiWebhookPayload {
  message: VapiWebhookPayloadMessage;
}

export type ActionItemType =
  (typeof ActionItemType)[keyof typeof ActionItemType];

export const ActionItemType = {
  map: "map",
  directions: "directions",
  call: "call",
  link: "link",
} as const;

export interface ActionItem {
  type: ActionItemType;
  label: string;
  query?: string;
  destination?: string;
  phone?: string;
  url?: string;
}

export interface VapiWebhookResponse {
  response: string;
  intent: string;
  isEmergency: boolean;
  language: string;
  actions?: ActionItem[];
}

export type QueryRequestLanguage =
  (typeof QueryRequestLanguage)[keyof typeof QueryRequestLanguage];

export const QueryRequestLanguage = {
  en: "en",
  hi: "hi",
  kn: "kn",
  auto: "auto",
} as const;

export interface QueryRequest {
  userId?: string;
  text: string;
  language?: QueryRequestLanguage;
}

export interface QueryResponse {
  response: string;
  intent: string;
  isEmergency: boolean;
  language: string;
  userId: string;
  memoryUsed: boolean;
  actions: ActionItem[];
}

export interface MemoryEntry {
  id: string;
  text: string;
  timestamp: string;
  language: string;
  intent: string;
}

export interface UserMemoryResponse {
  userId: string;
  memories: MemoryEntry[];
  count: number;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}