# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jeevan Voice API - Voice AI agent for low-literacy users
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
  qdrant: zod.string().optional(),
  openai: zod.string().optional(),
});

/**
 * Receives voice input events from Vapi and returns AI responses
 * @summary Vapi webhook handler
 */
export const VapiWebhookBody = zod.object({
  message: zod.object({
    type: zod.string(),
    call: zod
      .object({
        id: zod.string().optional(),
      })
      .optional(),
    transcript: zod.string().optional(),
    customer: zod
      .object({
        number: zod.string().optional(),
      })
      .optional(),
  }),
});

export const VapiWebhookResponse = zod.object({
  response: zod.string(),
  intent: zod.string(),
  isEmergency: zod.boolean(),
  language: zod.string(),
  actions: zod
    .array(
      zod.object({
        type: zod.enum(["map", "directions", "call", "link"]),
        label: zod.string(),
        query: zod.string().optional(),
        destination: zod.string().optional(),
        phone: zod.string().optional(),
        url: zod.string().optional(),
      }),
    )
    .optional(),
});

/**
 * Send a text query to the AI agent (for testing without voice)
 * @summary Text query endpoint
 */
export const queryTextBodyLanguageDefault = `auto`;

export const QueryTextBody = zod.object({
  userId: zod.string().optional(),
  text: zod.string(),
  language: zod
    .enum(["en", "hi", "kn", "auto"])
    .default(queryTextBodyLanguageDefault),
});

export const QueryTextResponse = zod.object({
  response: zod.string(),
  intent: zod.string(),
  isEmergency: zod.boolean(),
  language: zod.string(),
  userId: zod.string(),
  memoryUsed: zod.boolean(),
  actions: zod.array(
    zod.object({
      type: zod.enum(["map", "directions", "call", "link"]),
      label: zod.string(),
      query: zod.string().optional(),
      destination: zod.string().optional(),
      phone: zod.string().optional(),
      url: zod.string().optional(),
    }),
  ),
});

/**
 * Retrieve stored conversation memory for a user
 * @summary Get user memory
 */
export const GetUserMemoryParams = zod.object({
  userId: zod.coerce.string(),
});

export const GetUserMemoryResponse = zod.object({
  userId: zod.string(),
  memories: zod.array(
    zod.object({
      id: zod.string(),
      text: zod.string(),
      timestamp: zod.string(),
      language: zod.string(),
      intent: zod.string(),
    }),
  ),
  count: zod.number(),
});

/**
 * Clear all stored memory for a user
 * @summary Clear user memory
 */
export const ClearUserMemoryParams = zod.object({
  userId: zod.coerce.string(),
});

export const ClearUserMemoryResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});