# lib/api-zod/src/generated/types/actionItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jeevan Voice API - Voice AI agent for low-literacy users
 * OpenAPI spec version: 0.1.0
 */
import type { ActionItemType } from "./actionItemType";

export interface ActionItem {
  type: ActionItemType;
  label: string;
  query?: string;
  destination?: string;
  phone?: string;
  url?: string;
}