# lib/api-zod/src/generated/types/vapiWebhookResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jeevan Voice API - Voice AI agent for low-literacy users
 * OpenAPI spec version: 0.1.0
 */
import type { ActionItem } from "./actionItem";

export interface VapiWebhookResponse {
  response: string;
  intent: string;
  isEmergency: boolean;
  language: string;
  actions?: ActionItem[];
}